Refactored Fragment_Login to use HTTP queries on MongoDB instead of SQLite database
In this commit, I updated the Fragment_Login to use HTTP queries to communicate with a MongoDB database, replacing the previous implementation which used a SQLite database. This change should improve the app's scalability and performance by allowing for easier scaling of the database and reducing the amount of processing required on the client side.